backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1
redis==5.0.1
celery==5.3.6
aiogram==3.3.0
python-telegram-bot==20.7
boto3==1.34.34
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
websockets==12.0
prometheus-client==0.19.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
